The scale factor might not be an integer. (haog/gqe-python!163)
This is useful for CI, where we run with a scale factor that is less than one so that we can have better performance.

+++ gqe/benchmark/hardcoded/bindings/tpch_q13.py
@@ -29,7 +29,7 @@ class Q13GroupjoinBuildRelation(Relation):
     Q13 build relation for the customer table.
     """
 
-    def __init__(self, input: Relation, scale_factor: int):
+    def __init__(self, input: Relation, scale_factor: float):
         self.input = input
         self.scale_factor = scale_factor
